+++ lambdas/src/init/package.json
@@ -2,5 +2,17 @@
   "dependencies": {
     "assert": "^2.0.0",
     "uuid": "^3.4.0"
+  },
+  "devDependencies": {
+    "@types/assert": "^1.4.6",
+    "@types/aws-sdk": "^2.7.0",
+    "@types/node": "^13.7.1",
+    "@types/uuid": "^3.4.7",
+    "awesome-typescript-loader": "^5.2.1",
+    "clean-webpack-plugin": "^3.0.0",
+    "terser-webpack-plugin": "^2.3.4",
+    "typescript": "^3.7.5",
+    "webpack": "^4.41.6",
+    "webpack-cli": "^3.3.11"
   }
 }
