+++ lambdas/src/init/src/UrlPayload.ts
@@ -1,0 +1,4 @@
+export interface UrlPayload {
+    url: string,
+    metadata?: object
+}
