+++ lambdas/src/init/webpack.config.js
@@ -1,0 +1,40 @@
+const TerserPlugin = require("terser-webpack-plugin");
+const { CleanWebpackPlugin } = require("clean-webpack-plugin");
+
+module.exports = {
+  mode: "production",
+  target: "node",
+  performance: {
+    hints: false
+  },
+  entry: "./src/index.ts",
+  output: {
+    filename: "index.js",
+    library: "main",
+    libraryTarget: "commonjs2"
+  },
+
+  resolve: {
+    // Add '.ts' and '.tsx' as resolvable extensions.
+    extensions: [".webpack.js", ".web.js", ".ts", ".tsx", ".js"]
+  },
+
+  plugins: [new CleanWebpackPlugin()],
+  optimization: {
+    minimize: true,
+    minimizer: [
+      new TerserPlugin({
+        extractComments: false
+      })
+    ]
+  },
+
+  module: {
+    rules: [
+      {
+        test: /\.tsx?$/,
+        loader: "awesome-typescript-loader"
+      }
+    ]
+  }
+};
